back-end/requirements.txt
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
marshmallow==2.16.3
psycopg2-binary==2.7.6.1
python-dotenv==0.9.1
SQLAlchemy==1.2.14
Werkzeug==0.14.1
